+++ frontend/src/app/components/fabrica/edit-tareas-form/edit-tareas-form.component.ts
@@ -1,8 +1,8 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { Tarea } from '../../../interfaces/interfaces';
+import { Skill, Tarea } from '../../../interfaces/interfaces';
 import { TareasService } from '../../../services/tareas.service';
 import { ApiService } from '../../../services/api.service';
-import { finalize } from 'rxjs';
+import { Subscription, finalize } from 'rxjs';
 import { TareaImpl } from '../../../clases/tarea.class';
 
 @Component({
@@ -11,21 +11,47 @@ import { TareaImpl } from '../../../clases/tarea.class';
   styleUrl: './edit-tareas-form.component.css'
 })
 export class EditTareasFormComponent {
+  @Input() hard_skills: Skill[] = [];
+  @Input() soft_skills: Skill[] = [];
+
   @Output() close = new EventEmitter();
 
   cargando: boolean = false;
 
   @Input() tarea!: Tarea;
 
-  sector: string = "";
   nombre: string = "";
   duracion!: number;
   beneficio!: number;
   coste!: number;
   descripcion: string = "";
+  skills: number[] = [];
   subtask_dependencia: string = "";
 
+  tareas: Tarea[] = [];
+  private tareasSub?: Subscription;
+
   constructor(private apiService: ApiService, private tareasService: TareasService) { }
+
+  ngOnInit(): void {
+    this.tareasSub = this.tareasService.tareas$.subscribe(tareas => {
+      this.tareas = tareas;
+    });
+
+    this.nombre = this.tarea.nombre;
+    this.duracion = this.tarea.duracion;
+    this.beneficio = this.tarea.beneficio;
+    this.coste = this.tarea.coste;
+    this.descripcion = this.tarea.descripcion;
+    this.skills = this.tarea.skills;
+    this.subtask_dependencia = this.tarea.tareaPadre?.nombre != undefined ? "" + this.tarea.tareaPadre?.nombre : "Sin tarea padre";
+  }
+
+  ngOnDestroy(): void {
+    if (this.tareasSub) {
+      this.tareasSub.unsubscribe();
+    }
+  }
 
   cerrarModal(): void {
     if(!this.cargando) {
@@ -33,21 +59,12 @@ export class EditTareasFormComponent {
     }
   }
 
-  ngOnInit(): void {
-    this.nombre = this.tarea.nombre;
-    this.duracion = this.tarea.duracion;
-    this.beneficio = this.tarea.beneficio;
-    this.coste = this.tarea.coste;
-    this.descripcion = this.tarea.nombre;
-    this.subtask_dependencia = this.tarea.getTareaPadre()?.id != undefined ? "" + this.tarea.getTareaPadre()?.id : "";
-  }
-
   modificarTarea(): void {
     if(!this.cargando){
       console.log("Modificando la tarea...");
       this.cargando = true;
 
-      this.apiService.modificarTarea(this.tarea.id, this.tarea.nombre, this.tarea.duracion, this.tarea.beneficio, this.tarea.descripcion).pipe(
+      this.apiService.modificarTarea(this.tarea.id, this.nombre, this.duracion, this.beneficio, this.coste, this.descripcion, this.skills).pipe(
         finalize(() => {
           this.cargando = false; 
           this.cerrarModal();
@@ -64,11 +81,13 @@ export class EditTareasFormComponent {
             const duracion = response.subtask[2];
             const beneficio = response.subtask[3];
             const descripcion = response.subtask[4];
-            const coste = 0;
+            //const fabrica_id = response.tarea[5];
+            const coste = response.subtask[6];
+            const skills = response.subtask[7];
 
             //Si tenemos todos los datos añadimos la tarea
-            if(id != undefined && nombre != undefined && duracion != undefined && beneficio != undefined && descripcion != undefined) {
-              this.tareasService.actualizarTarea(new TareaImpl(id, nombre, 0, duracion, beneficio, coste, 0, descripcion));
+            if(id != undefined && nombre != undefined && duracion != undefined && beneficio != undefined && descripcion != undefined && coste != undefined && skills != undefined) {
+              this.tareasService.modificarTarea(id, nombre, duracion, beneficio, coste, descripcion, skills);
             } else {
               console.log("Omitiendo la modificacion de la tarea por falta de datos...");
             }
@@ -84,22 +103,38 @@ export class EditTareasFormComponent {
   }
 
   borrarTarea() {
-    if(confirm("¿Estás seguro que deseas eliminar la tarea?")) {
-      console.log("Eliminando la tarea...");
-    
-      this.apiService.eliminarTarea(this.tarea.id).pipe(
-        finalize(() => {
-          console.log("Fin de eliminar tarea.");
-        })
-      ).subscribe({
-        next: (response) => {
-          console.log("Respuesta: ", response);
-          this.tareasService.eliminarTarea(this.tarea.id);
-        },
-        error: (error) => {
-          alert("Error: " + error); 
-        }
-      });
+    if(!this.cargando){
+      if(confirm("¿Estás seguro que deseas eliminar la tarea?")) {
+        console.log("Eliminando la tarea...");
+        this.cargando = true;
+      
+        this.apiService.eliminarTarea(this.tarea.id).pipe(
+          finalize(() => {
+            console.log("Fin de eliminar tarea.");
+            this.cargando = false;
+          })
+        ).subscribe({
+          next: (response) => {
+            console.log("Respuesta: ", response);
+            this.tareasService.eliminarTarea(this.tarea.id);
+            this.close.emit();
+          },
+          error: (error) => {
+            alert("Error: " + error); 
+          }
+        });
+      }
+    }
+  }
+
+  checkSkill(id: number) {
+    const index = this.skills.indexOf(id);
+    if (index !== -1) {
+      // Si el valor ya está presente, lo quitamos
+      this.skills.splice(index, 1);
+    } else {
+      // Si el valor no está presente, lo agregamos
+      this.skills.push(id);
     }
   }
 }
